:test_tube: Adding test scenarios for the operations

+++ app/src/main/java/com/victor/composecalculator/model/UiEvent.kt
@@ -1,7 +1,7 @@
 package com.victor.composecalculator.model
 
 sealed class UiEvent {
-    data class TypeNumber(val number: Int): UiEvent()
+    data class TypeNumber(val number: Double): UiEvent()
     data class AddOperation(val operation: Operation): UiEvent()
     object AddDecimalCharacter: UiEvent()
     object CalculateOperation: UiEvent()
